src/requirements.txt
# Core RAG Dependencies
chromadb
openai
langchain
langchain-openai
langchain-community
langchain-google-genai
langgraph
google-generativeai==0.3.2

# Document Processing
PyPDF2
pdfplumber
camelot-py
pandas
PyMuPDF
pillow
pytesseract

# Additional dependencies for camelot
opencv-python

# Embeddings and ML
sentence-transformers
numpy
scikit-learn

# Web Access
requests

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml

# API and Web Framework (for future milestones)
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Testing
pytest
pytest-asyncio

# Utilities
tqdm
click
colorama

# Development Tools
black
flake8
isort

# Additional dependencies
python-multipart==0.0.6
pathlib==1.0.1
typing-extensions>=4.9.0

selenium
webdriver-manager 
beautifulsoup4
undetected-chromedriver
html2text
pdfminer.six

google-api-python-client
google_auth
google-auth-oauthlib
google-auth-httplib2
